requirements.txt
streamlit
serpapi
pandas
st-gsheets-connection
langchain-groq
langchain-community
python-dotenv